Added more pickups, changed where they spawn, added more options to TextDisplay script, implemented ScoreMultiplier reset on exiting playfield

+++ Assets/Scripts/PlayerScripts/Player.cs
@@ -41,7 +41,7 @@ namespace PlayerScripts
             PlayerStatus.ChangeBombs(-1);
         }
 
-        private void CheckBombs(int bombs)
+        private void CheckBombs(float bombs)
         {
             _hasBombs = bombs > 0;
         }
